+++ zk-token-sdk/src/errors.rs
@@ -1,13 +1,42 @@
 //! Errors related to proving and verifying proofs.
-use {
-    crate::{
-        encryption::elgamal::ElGamalError,
-        range_proof::errors::{RangeProofGenerationError, RangeProofVerificationError},
-        sigma_proofs::errors::*,
-    },
-    thiserror::Error,
+#[cfg(not(target_os = "solana"))]
+use crate::{
+    range_proof::errors::{RangeProofGenerationError, RangeProofVerificationError},
+    sigma_proofs::errors::*,
 };
+use thiserror::Error;
 
+#[derive(Error, Clone, Debug, Eq, PartialEq)]
+pub enum AuthenticatedEncryptionError {
+    #[error("key derivation method not supported")]
+    DerivationMethodNotSupported,
+    #[error("seed length too short for derivation")]
+    SeedLengthTooShort,
+    #[error("seed length too long for derivation")]
+    SeedLengthTooLong,
+    #[error("failed to deserialize")]
+    Deserialization,
+}
+
+#[derive(Error, Clone, Debug, Eq, PartialEq)]
+pub enum ElGamalError {
+    #[error("key derivation method not supported")]
+    DerivationMethodNotSupported,
+    #[error("seed length too short for derivation")]
+    SeedLengthTooShort,
+    #[error("seed length too long for derivation")]
+    SeedLengthTooLong,
+    #[error("failed to deserialize ciphertext")]
+    CiphertextDeserialization,
+    #[error("failed to deserialize public key")]
+    PubkeyDeserialization,
+    #[error("failed to deserialize keypair")]
+    KeypairDeserialization,
+    #[error("failed to deserialize secret key")]
+    SecretKeyDeserialization,
+}
+
+#[cfg(not(target_os = "solana"))]
 #[derive(Error, Clone, Debug, Eq, PartialEq)]
 pub enum ProofGenerationError {
     #[error("not enough funds in account")]
@@ -26,6 +55,7 @@ pub enum ProofGenerationError {
     ProofLength,
 }
 
+#[cfg(not(target_os = "solana"))]
 #[derive(Error, Clone, Debug, Eq, PartialEq)]
 pub enum ProofVerificationError {
     #[error("range proof verification failed")]
@@ -51,35 +81,42 @@ pub enum SigmaProofType {
     PubkeyValidityProof,
 }
 
+#[cfg(not(target_os = "solana"))]
 #[derive(Error, Clone, Debug, Eq, PartialEq)]
 pub enum TranscriptError {
     #[error("point is the identity")]
     ValidationError,
 }
 
+#[cfg(not(target_os = "solana"))]
 impl From<EqualityProofVerificationError> for ProofVerificationError {
     fn from(err: EqualityProofVerificationError) -> Self {
         Self::SigmaProof(SigmaProofType::EqualityProof, err.0)
     }
 }
 
+#[cfg(not(target_os = "solana"))]
 impl From<FeeSigmaProofVerificationError> for ProofVerificationError {
     fn from(err: FeeSigmaProofVerificationError) -> Self {
         Self::SigmaProof(SigmaProofType::FeeSigmaProof, err.0)
     }
 }
 
+#[cfg(not(target_os = "solana"))]
 impl From<ZeroBalanceProofVerificationError> for ProofVerificationError {
     fn from(err: ZeroBalanceProofVerificationError) -> Self {
         Self::SigmaProof(SigmaProofType::ZeroBalanceProof, err.0)
     }
 }
+
+#[cfg(not(target_os = "solana"))]
 impl From<ValidityProofVerificationError> for ProofVerificationError {
     fn from(err: ValidityProofVerificationError) -> Self {
         Self::SigmaProof(SigmaProofType::ValidityProof, err.0)
     }
 }
 
+#[cfg(not(target_os = "solana"))]
 impl From<PubkeyValidityProofVerificationError> for ProofVerificationError {
     fn from(err: PubkeyValidityProofVerificationError) -> Self {
         Self::SigmaProof(SigmaProofType::PubkeyValidityProof, err.0)
